Remade files and fixed bug
I remade the files truncating the data to the smallest data types in the built in types from numpy. I made a very simple mistake previously when making the files and wrote the ROI object not the ROI data to the json file. That was blowing up the data massively.